backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic-settings>=2.3.0
openai>=1.40.0
python-dotenv>=1.0.1
